+++ Source/Examples/WPF.SharpDX/ExampleBrowser/Workitems/Workitem1349/MainWindow.xaml.cs
@@ -1,0 +1,26 @@
+﻿// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="MainWindow.xaml.cs" company="Helix Toolkit">
+//   Copyright (c) 2014 Helix Toolkit contributors
+// </copyright>
+// <summary>
+//   Interaction logic for MainWindow.xaml
+// </summary>
+// --------------------------------------------------------------------------------------------------------------------
+
+namespace Workitem1349
+{
+    using System.Windows;
+    using ExampleBrowser;
+
+    /// <summary>
+    /// Interaction logic for MainWindow.xaml
+    /// </summary>
+    [Example("Issue 1349", "SharpDX: BillboardText3D is not shown until origin is moved into sight.")]
+    public partial class MainWindow : Window
+    {
+        public MainWindow()
+        {
+            this.InitializeComponent();
+        }
+    }
+}