+++ Source/HelixToolkit.Wpf.SharpDX.Tests/Elements3D/MeshGeometryModel3DTests.cs
@@ -1,28 +1,28 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
 // <copyright file="MeshGeometryModel3DTests.cs" company="Helix Toolkit">
-//   Copyright (c) 2014 Helix Toolkit contributors
+//   Copyright (c) 2020 Helix Toolkit contributors
 // </copyright>
-// <summary>
-//   Test for pull request #54.
-//   Fix IndexOutOfRange exception in CreateDefaultVertexArray.
-// </summary>
 // --------------------------------------------------------------------------------------------------------------------
 
 using System;
+using System.Collections.Generic;
 using HelixToolkit.Wpf.SharpDX.Tests.Controls;
 using NUnit.Framework;
 using System.IO;
+using System.Threading;
+using SharpDX;
 
 namespace HelixToolkit.Wpf.SharpDX.Tests.Elements3D
 {
     [TestFixture]
+    [Apartment(ApartmentState.STA)]
     class MeshGeometryModel3DTests
     {
         /// <summary>
         /// Test for pull request #54.
         /// Fix IndexOutOfRange exception in CreateDefaultVertexArray.
         /// </summary>
-        [Test, STAThread]
+        [Test]
         public void CreateDefaultVertexArrayForTriangle()
         {
             var reader = new ObjReader();
@@ -37,6 +37,28 @@ namespace HelixToolkit.Wpf.SharpDX.Tests.Elements3D
             model.SceneNode.Attach(canvas.RenderHost);
 
             Assert.AreEqual(true, model.IsAttached);
+        }
+
+        private MeshGeometryModel3D GetGeometryModel3D()
+        {
+            var meshBuilder = new MeshBuilder();
+            meshBuilder.AddBox(new Vector3(0f), 1, 1, 1);
+            return new MeshGeometryModel3D()
+            {
+                Geometry = meshBuilder.ToMesh(),
+            };
+        }
+
+        [Test]
+        public void HitTestShouldReturnOnePointOnFrontOfCubeWithNoCuttingPlanes()
+        {
+            var viewport = new Viewport3DX();
+            var ray = new Ray(new Vector3(2f, 0f, 0f), new Vector3(-1, 0, 0));
+            var hits = new List<HitTestResult>();
+            var geometryModel3D = GetGeometryModel3D();
+            geometryModel3D.HitTest(viewport.RenderContext, ray, ref hits);
+            Assert.AreEqual(1, hits.Count);
+            Assert.AreEqual(new Vector3(0.5f, 0, 0), hits[0].PointHit);
         }
     }
 }