+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [2.12.0] - 2020-05-25
+### Added
+1. Support Vertex Color blending for Phong/PBR/Diffuse materials with new `VertexColorBlendingFactor` property. (WPF.SharpDX/UWP/Core)
+
+### Fixed
+1. Fix billboardText not being call to initialize. (WPF.SharpDX/UWP/Core)
+2. Empty BillboardSingleText3D Causes Crash. (WPF.SharpDX/UWP/Core)
+3. Argument exception when using View.RenderTargetBitmap. (WPF.SharpDX/UWP/Core)
+
 ## [2.11.0] - 2020-02-08
 ### Added
 1. Add IsTopBottomViewOrientedToFrontBack property to view cube #1263. (WPF)